feat(standard_layout): Move main package to cmd/
